Move some packages & fix nicknames being set randomly
