+++ CarbonChat-Bukkit/src/main/java/net/draycia/carbon/bukkit/CarbonChatBukkit.java
@@ -5,8 +5,8 @@ import cloud.commandframework.execution.CommandExecutionCoordinator;
 import cloud.commandframework.paper.PaperCommandManager;
 import net.draycia.carbon.api.channels.ChatChannel;
 import net.draycia.carbon.bukkit.config.CarbonBukkitSettings;
-import net.draycia.carbon.bukkit.listeners.events.BukkitChatListener;
-import net.draycia.carbon.bukkit.listeners.events.ItemLinkHandler;
+import net.draycia.carbon.bukkit.listeners.BukkitChatListener;
+import net.draycia.carbon.bukkit.listeners.ItemLinkHandler;
 import net.draycia.carbon.bukkit.messaging.BungeeMessageService;
 import net.draycia.carbon.bukkit.users.BukkitConsoleUser;
 import net.draycia.carbon.bukkit.users.BukkitPlayerUser;
@@ -18,7 +18,7 @@ import net.draycia.carbon.api.CarbonChatProvider;
 import net.draycia.carbon.api.config.CarbonTranslations;
 import net.draycia.carbon.api.adventure.MessageProcessor;
 import net.draycia.carbon.api.channels.ChannelRegistry;
-import net.draycia.carbon.api.commands.settings.CommandSettingsRegistry;
+import net.draycia.carbon.api.commands.CommandSettingsRegistry;
 import net.draycia.carbon.api.config.CarbonSettings;
 import net.draycia.carbon.api.config.ChannelSettings;
 import net.draycia.carbon.api.config.MessagingType;
@@ -32,20 +32,20 @@ import net.draycia.carbon.common.config.ConfigLoader;
 import net.draycia.carbon.common.messaging.EmptyMessageService;
 import net.draycia.carbon.api.config.SQLCredentials;
 import net.draycia.carbon.common.messaging.RedisMessageService;
-import net.draycia.carbon.common.listeners.events.FilterHandler;
-import net.draycia.carbon.common.listeners.events.CapsHandler;
-import net.draycia.carbon.common.listeners.events.CustomPlaceholderHandler;
-import net.draycia.carbon.common.listeners.events.IgnoredPlayerHandler;
-import net.draycia.carbon.common.listeners.events.LegacyFormatHandler;
-import net.draycia.carbon.common.listeners.events.MuteHandler;
-import net.draycia.carbon.common.listeners.events.PingHandler;
-import net.draycia.carbon.bukkit.listeners.events.PlaceholderHandler;
-import net.draycia.carbon.common.listeners.events.PlayerJoinListener;
-import net.draycia.carbon.bukkit.listeners.events.RelationalPlaceholderHandler;
-import net.draycia.carbon.common.listeners.events.ShadowMuteHandler;
-import net.draycia.carbon.common.listeners.events.UrlLinkHandler;
-import net.draycia.carbon.common.listeners.events.UserFormattingHandler;
-import net.draycia.carbon.common.listeners.events.WhisperPingHandler;
+import net.draycia.carbon.common.listeners.FilterHandler;
+import net.draycia.carbon.common.listeners.CapsHandler;
+import net.draycia.carbon.common.listeners.CustomPlaceholderHandler;
+import net.draycia.carbon.common.listeners.IgnoredPlayerHandler;
+import net.draycia.carbon.common.listeners.LegacyFormatHandler;
+import net.draycia.carbon.common.listeners.MuteHandler;
+import net.draycia.carbon.common.listeners.PingHandler;
+import net.draycia.carbon.bukkit.listeners.PlaceholderHandler;
+import net.draycia.carbon.common.listeners.PlayerJoinListener;
+import net.draycia.carbon.bukkit.listeners.RelationalPlaceholderHandler;
+import net.draycia.carbon.common.listeners.ShadowMuteHandler;
+import net.draycia.carbon.common.listeners.UrlLinkHandler;
+import net.draycia.carbon.common.listeners.UserFormattingHandler;
+import net.draycia.carbon.common.listeners.WhisperPingHandler;
 import net.draycia.carbon.common.adventure.AdventureManager;
 import net.draycia.carbon.common.commands.misc.CommandRegistrar;
 import net.draycia.carbon.common.messaging.MessageManager;
