+++ CarbonChat-Bukkit/src/main/java/net/draycia/carbon/bukkit/listeners/PlaceholderHandler.java
@@ -1,4 +1,4 @@
-package net.draycia.carbon.bukkit.listeners.events;
+package net.draycia.carbon.bukkit.listeners;
 
 import me.clip.placeholderapi.PlaceholderAPI;
 import net.draycia.carbon.api.events.misc.CarbonEvents;
