+++ CarbonChat-Common/src/main/java/net/draycia/carbon/common/commands/ChannelListCommand.java
@@ -6,7 +6,7 @@ import net.draycia.carbon.api.CarbonChat;
 import net.draycia.carbon.api.CarbonChatProvider;
 import net.draycia.carbon.api.channels.ChatChannel;
 import net.draycia.carbon.api.users.CarbonUser;
-import net.draycia.carbon.api.commands.settings.CommandSettings;
+import net.draycia.carbon.api.commands.CommandSettings;
 import net.draycia.carbon.api.users.PlayerUser;
 import net.draycia.carbon.common.commands.arguments.PlayerUserArgument;
 import net.kyori.adventure.identity.Identity;
