+++ CarbonChat-Common/src/main/java/net/draycia/carbon/common/commands/ChatReloadCommand.java
@@ -4,7 +4,7 @@ import cloud.commandframework.CommandManager;
 import cloud.commandframework.context.CommandContext;
 import net.draycia.carbon.api.CarbonChat;
 import net.draycia.carbon.api.CarbonChatProvider;
-import net.draycia.carbon.api.commands.settings.CommandSettings;
+import net.draycia.carbon.api.commands.CommandSettings;
 import net.draycia.carbon.api.users.CarbonUser;
 import net.kyori.adventure.identity.Identity;
 import net.kyori.adventure.text.Component;
