+++ CarbonChat-Common/src/main/java/net/draycia/carbon/common/commands/PingOptionsCommand.java
@@ -7,7 +7,7 @@ import cloud.commandframework.arguments.standard.FloatArgument;
 import cloud.commandframework.context.CommandContext;
 import net.draycia.carbon.api.CarbonChat;
 import net.draycia.carbon.api.CarbonChatProvider;
-import net.draycia.carbon.api.commands.settings.CommandSettings;
+import net.draycia.carbon.api.commands.CommandSettings;
 import net.draycia.carbon.api.users.CarbonUser;
 import net.draycia.carbon.api.users.PlayerUser;
 import net.kyori.adventure.key.Key;
