+++ CarbonChat-Common/src/main/java/net/draycia/carbon/common/listeners/LegacyFormatHandler.java
@@ -1,4 +1,4 @@
-package net.draycia.carbon.common.listeners.events;
+package net.draycia.carbon.common.listeners;
 
 import net.draycia.carbon.api.events.PreChatFormatEvent;
 import net.draycia.carbon.api.events.misc.CarbonEvents;
